backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==0.28.1
python-dotenv==1.0.0
# Use a Pydantic version with wheels for Python 3.13 to avoid building Rust
pydantic==2.8.2
python-multipart==0.0.6
flask==3.0.0
numpy==1.26.4
requests==2.31.0